commit before pull

+++ routes/products.js
@@ -1,26 +1,16 @@
 const express = require('express')
 const router = express.Router()
 const Product = require('../models/Product')
-const User = require('../models/User')
+// const User = require('../models/User')
 const path = require('path')
 const { ensureLoggedIn, ensureLoggedOut } = require('connect-ensure-login')
 const multer = require('multer')
 const destination = path.join(__dirname, "../public/avatar/")
 const upload = multer({dest : destination})
 
-router.get('/products', ensureLoggedIn(), (req, res, next) => {
-  const user = req.params
-  const userId = req.user._id
-  Product.find({'user_id':userId})
-  .then( response => {
-    res.render('products/index', {title:'Lista de productos',products: response, user:user})
-  })
-  .catch( err => next(err) )
-})
 
-// CREATE
 router.get('/products/new', ensureLoggedIn(), (req, res, next) => {
-  res.render('products/new',{title:'Add a product'})
+  res.render('products/new',{subtitle:'Add a product'})
 })
 
 router.post('/products/new', upload.single('avatar'), (req,res,next) =>{
@@ -33,31 +23,27 @@ router.post('/products/new', upload.single('avatar'), (req,res,next) =>{
   }
   const newProduct = new Product(productInfo)
   newProduct.save()
-  .then( response => { res.redirect('/products') })
+  .then( response => { res.redirect(`/profile/${req.user._id}`) })
   .catch( err => next(err) )
 })
 
-// RETRIEVE
 router.get('/products/:id', (req, res, next) => {
   const productId = req.params.id
-
   Product.findById(productId)
-  .then( response => {
-    res.render('products/details',{title: 'Detalle', product: response})
-  })
+  .then( response => { res.render('products/details',{product: response}) })
   .catch( err => next(err) )
-
 })
 
-
-// UPDATE
-router.get('/products/:id/edit', (req, res, next) => {
+router.get('/products/:id/edit', ensureLoggedIn(), (req, res, next) => {
   const productId = req.params.id
-  Product.findById(productId)
-  .then( response => {
-    res.render('products/edit', { title:'Modify form', product: response, productId })
-  })
-  .catch( err => next(err))
+    Product.findById(productId)
+    .then( response => {
+      if( response.user_id != req.user._id){
+        return res.redirect('/')
+      }
+      return res.render('products/edit', { subtitle:'Edit product', product: response, productId })
+    })
+    .catch( err => next(err))
 })
 
 router.post('/products/:id/edit', upload.single('avatar'), (req, res, next) => {
@@ -68,18 +54,21 @@ router.post('/products/:id/edit', upload.single('avatar'), (req, res, next) => {
         description: req.body.description,
         avatar: `/avatar/${req.file.filename}`
   }
-  Product.findByIdAndUpdate(productId, updates, (err, product) => {
-    if (err){ return next(err) }
-    return res.redirect(`/products/${productId}`)
-  })
+  Product.findByIdAndUpdate(productId, updates)
+  .then( response => { res.redirect(`/products/${productId}`) })
+  .catch( err => next(err) )
 })
 
-// DELETE
-router.get('/products/:id/delete', (req, res, next) => {
+router.get('/products/:id/delete', ensureLoggedIn(), (req, res, next) => {
   const productId = req.params.id
-  Product.findByIdAndRemove(productId)
-  .then( response => res.redirect('/products'))
-  .catch( err => next(err))
+  if( req.params.user_id == req.user._id){
+    Product.findByIdAndRemove(productId)
+    .then( response => res.redirect(`/profile/${req.user._id}`))
+    .catch( err => next(err))
+  }else{
+    return res.redirect('/')
+  }
+
 })
 
 module.exports = router
